requirements_old.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
billiard==4.1.0
bitarray==2.8.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
Django==4.2
django-appconf==1.0.5
django-ckeditor==6.7.0
django-cors-headers==3.14.0
django-js-asset==2.1.0
django-nested-admin==4.0.2
django-phonenumber-field==7.0.2
django-select2==8.1.2
djangochannelsrestframework==1.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.3.2
msgpack==1.0.5
multidict==6.0.4
openai==0.27.6
pdfreader==0.1.12
phonenumbers==8.13.10
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.6.0
pyOpenSSL==23.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-monkey-business==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.5.4
requests==2.29.0
service-identity==21.1.0
sib-api-v3-sdk==7.6.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.65.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.22.0
uvloop==0.17.0
uWSGI==2.0.21
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.2
yarl==1.9.2
zope.interface==6.0
